refactor: Refactor user services and repository
- Move user services to the services package
- Add repository interface to user services
- Update user services constructor to accept repository
- Implement GetUserByID method in MySQLUserRepository

Refactoring the user services and repository to improve code organization and separation of concerns. The user services are now located in the services package, and a repository interface has been added to the services. The constructor for the user services now accepts a repository as a parameter. Additionally, the GetUserByID method has been implemented in the MySQLUserRepository.